E-Commerce-Chatbot-AstraDB-Langchain/requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
flask
setuptools
pandas

-e .